Refactor ExtraYAxisLabels.
